Move TimestampContext
